+++ src/cluster/cluster_configuration.go
@@ -469,6 +469,7 @@ func (self *ClusterConfiguration) SaveClusterAdmin(u *ClusterAdmin) {
 		return
 	}
 	self.clusterAdmins[u.GetName()] = u
+	u.ChangePassword(u.Hash)
 }
 
 type SavedConfiguration struct {